Move mapping utilities to parallel_map.py
To avoid some circular dependencies, a few thing had to move to function_utils as well
