+++ modal/_asgi.py
@@ -4,12 +4,12 @@ from typing import Any, AsyncGenerator, Callable, Dict, Optional, cast
 
 import aiohttp
 
+from . import current_function_call_id
 from ._utils.async_utils import TaskContext
 from ._utils.blob_utils import MAX_OBJECT_SIZE_BYTES
 from .config import logger
 from .exception import ExecutionError, InvalidError
 from .experimental import stop_fetching_inputs
-from .functions import current_function_call_id
 
 FIRST_MESSAGE_TIMEOUT_SECONDS = 5.0
