+++ modal/functions.py
@@ -4,7 +4,6 @@ import inspect
 import time
 import typing
 import warnings
-from contextvars import ContextVar
 from dataclasses import dataclass
 from pathlib import PurePosixPath
 from typing import (
@@ -20,18 +19,15 @@ from typing import (
     List,
     Optional,
     Sequence,
-    Set,
     Sized,
-    Tuple,
     Type,
     Union,
 )
 
-from aiostream import pipe, stream
+from aiostream import stream
 from google.protobuf.message import Message
 from grpclib import GRPCError, Status
 from synchronicity.combined_types import MethodWithAio
-from synchronicity.exceptions import UserCodeException
 
 from modal_proto import api_grpc, api_pb2
 
@@ -40,34 +36,34 @@ from ._location import parse_cloud_provider
 from ._output import OutputManager
 from ._pty import get_pty_info
 from ._resolver import Resolver
-from ._serialization import deserialize, deserialize_data_format, serialize
-from ._traceback import append_modal_tb
+from ._serialization import serialize
 from ._utils.async_utils import (
     AsyncOrSyncIteratable,
-    queue_batch_iterator,
     synchronize_api,
     synchronizer,
     warn_if_generator_is_not_consumed,
 )
-from ._utils.blob_utils import (
-    BLOB_MAX_PARALLELISM,
-    MAX_OBJECT_SIZE_BYTES,
-    blob_download,
-    blob_upload,
+from ._utils.function_utils import (
+    ATTEMPT_TIMEOUT_GRACE_PERIOD,
+    OUTPUTS_TIMEOUT,
+    FunctionInfo,
+    _create_input,
+    _process_result,
+    _stream_function_call_data,
+    current_input_id,
+    get_referred_objects,
+    is_async,
 )
-from ._utils.function_utils import FunctionInfo, _stream_function_call_data, get_referred_objects, is_async
 from ._utils.grpc_utils import retry_transient_errors
 from ._utils.mount_utils import validate_mount_points, validate_volumes
 from .call_graph import InputInfo, _reconstruct_call_graph
 from .client import _Client
 from .cloud_bucket_mount import _CloudBucketMount, cloud_bucket_mounts_to_proto
-from .config import config, logger
+from .config import config
 from .exception import (
     ExecutionError,
-    FunctionTimeoutError,
     InvalidError,
     NotFoundError,
-    RemoteError,
     deprecation_warning,
 )
 from .gpu import GPU_T, parse_gpu_config
@@ -75,6 +71,7 @@ from .image import _Image
 from .mount import _get_client_mount, _Mount
 from .network_file_system import _NetworkFileSystem, network_file_system_mount_protos
 from .object import Object, _get_environment_name, _Object, live_method, live_method_gen
+from .parallel_map import SynchronizedQueue, _map_invocation, _SynchronizedQueue
 from .proxy import _Proxy
 from .retries import Retries
 from .schedule import Schedule
@@ -82,131 +79,8 @@ from .scheduler_placement import SchedulerPlacement
 from .secret import _Secret
 from .volume import _Volume
 
-OUTPUTS_TIMEOUT = 55.0  # seconds
-ATTEMPT_TIMEOUT_GRACE_PERIOD = 5  # seconds
-
-
 if TYPE_CHECKING:
     import modal.stub
-
-
-class _SynchronizedQueue:
-    """mdmd:hidden"""
-
-    # small wrapper around asyncio.Queue to make it cross-thread compatible through synchronicity
-    async def init(self):
-        # in Python 3.8 the asyncio.Queue is bound to the event loop on creation
-        # so it needs to be created in a synchronicity-wrapped init method
-        self.q = asyncio.Queue()
-
-    @synchronizer.no_io_translation
-    async def put(self, item):
-        await self.q.put(item)
-
-    @synchronizer.no_io_translation
-    async def get(self):
-        return await self.q.get()
-
-
-SynchronizedQueue = synchronize_api(_SynchronizedQueue)
-
-
-def exc_with_hints(exc: BaseException):
-    """mdmd:hidden"""
-    if isinstance(exc, ImportError) and exc.msg == "attempted relative import with no known parent package":
-        exc.msg += """\n
-HINT: For relative imports to work, you might need to run your modal app as a module. Try:
-- `python -m my_pkg.my_app` instead of `python my_pkg/my_app.py`
-- `modal deploy my_pkg.my_app` instead of `modal deploy my_pkg/my_app.py`
-"""
-    elif isinstance(
-        exc, RuntimeError
-    ) and "CUDA error: no kernel image is available for execution on the device" in str(exc):
-        msg = (
-            exc.args[0]
-            + """\n
-HINT: This error usually indicates an outdated CUDA version. Older versions of torch (<=1.12)
-come with CUDA 10.2 by default. If pinning to an older torch version, you can specify a CUDA version
-manually, for example:
--  image.pip_install("torch==1.12.1+cu116", find_links="https://download.pytorch.org/whl/torch_stable.html")
-"""
-        )
-        exc.args = (msg,)
-
-    return exc
-
-
-async def _process_result(result: api_pb2.GenericResult, data_format: int, stub, client=None):
-    if result.WhichOneof("data_oneof") == "data_blob_id":
-        data = await blob_download(result.data_blob_id, stub)
-    else:
-        data = result.data
-
-    if result.status == api_pb2.GenericResult.GENERIC_STATUS_TIMEOUT:
-        raise FunctionTimeoutError(result.exception)
-    elif result.status != api_pb2.GenericResult.GENERIC_STATUS_SUCCESS:
-        if data:
-            try:
-                exc = deserialize(data, client)
-            except Exception as deser_exc:
-                raise ExecutionError(
-                    "Could not deserialize remote exception due to local error:\n"
-                    + f"{deser_exc}\n"
-                    + "This can happen if your local environment does not have the remote exception definitions.\n"
-                    + "Here is the remote traceback:\n"
-                    + f"{result.traceback}"
-                )
-            if not isinstance(exc, BaseException):
-                raise ExecutionError(f"Got remote exception of incorrect type {type(exc)}")
-
-            if result.serialized_tb:
-                try:
-                    tb_dict = deserialize(result.serialized_tb, client)
-                    line_cache = deserialize(result.tb_line_cache, client)
-                    append_modal_tb(exc, tb_dict, line_cache)
-                except Exception:
-                    pass
-            uc_exc = UserCodeException(exc_with_hints(exc))
-            raise uc_exc
-        raise RemoteError(result.exception)
-
-    try:
-        return deserialize_data_format(data, data_format, client)
-    except ModuleNotFoundError as deser_exc:
-        raise ExecutionError(
-            "Could not deserialize result due to error:\n"
-            + f"{deser_exc}\n"
-            + "This can happen if your local environment does not have a module that was used to construct the result. \n"
-        )
-
-
-async def _create_input(args, kwargs, client, idx: Optional[int] = None) -> api_pb2.FunctionPutInputsItem:
-    """Serialize function arguments and create a FunctionInput protobuf,
-    uploading to blob storage if needed.
-    """
-    if idx is None:
-        idx = 0
-
-    args_serialized = serialize((args, kwargs))
-
-    if len(args_serialized) > MAX_OBJECT_SIZE_BYTES:
-        args_blob_id = await blob_upload(args_serialized, client.stub)
-
-        return api_pb2.FunctionPutInputsItem(
-            input=api_pb2.FunctionInput(args_blob_id=args_blob_id, data_format=api_pb2.DATA_FORMAT_PICKLE),
-            idx=idx,
-        )
-    else:
-        return api_pb2.FunctionPutInputsItem(
-            input=api_pb2.FunctionInput(args=args_serialized, data_format=api_pb2.DATA_FORMAT_PICKLE),
-            idx=idx,
-        )
-
-
-@dataclass
-class _OutputValue:
-    # box class for distinguishing None results from non-existing/None markers
-    value: Any
 
 
 class _Invocation:
@@ -319,185 +193,6 @@ class _Invocation:
                 # and produces less data in the second run than what was already sent.
                 if items_total is not None and items_received >= items_total:
                     break
-
-
-MAP_INVOCATION_CHUNK_SIZE = 49
-
-
-async def _map_invocation(
-    function_id: str,
-    raw_input_queue: _SynchronizedQueue,
-    client: _Client,
-    order_outputs: bool,
-    return_exceptions: bool,
-    count_update_callback: Optional[Callable[[int, int], None]],
-):
-    assert client.stub
-    request = api_pb2.FunctionMapRequest(
-        function_id=function_id,
-        parent_input_id=current_input_id() or "",
-        function_call_type=api_pb2.FUNCTION_CALL_TYPE_MAP,
-        return_exceptions=return_exceptions,
-    )
-    response = await retry_transient_errors(client.stub.FunctionMap, request)
-
-    function_call_id = response.function_call_id
-
-    have_all_inputs = False
-    num_inputs = 0
-    num_outputs = 0
-
-    def count_update():
-        if count_update_callback is not None:
-            count_update_callback(num_outputs, num_inputs)
-
-    pending_outputs: Dict[str, int] = {}  # Map input_id -> next expected gen_index value
-    completed_outputs: Set[str] = set()  # Set of input_ids whose outputs are complete (expecting no more values)
-
-    input_queue: asyncio.Queue = asyncio.Queue()
-
-    async def create_input(argskwargs):
-        nonlocal num_inputs
-        idx = num_inputs
-        num_inputs += 1
-        (args, kwargs) = argskwargs
-        return await _create_input(args, kwargs, client, idx)
-
-    async def input_iter():
-        while 1:
-            raw_input = await raw_input_queue.get()
-            if raw_input is None:  # end of input sentinel
-                return
-            yield raw_input  # args, kwargs
-
-    async def drain_input_generator():
-        # Parallelize uploading blobs
-        proto_input_stream = stream.iterate(input_iter()) | pipe.map(
-            create_input,  # type: ignore[reportArgumentType]
-            ordered=True,
-            task_limit=BLOB_MAX_PARALLELISM,
-        )
-        async with proto_input_stream.stream() as streamer:
-            async for item in streamer:
-                await input_queue.put(item)
-
-        # close queue iterator
-        await input_queue.put(None)
-        yield
-
-    async def pump_inputs():
-        assert client.stub
-        nonlocal have_all_inputs, num_inputs
-        async for items in queue_batch_iterator(input_queue, MAP_INVOCATION_CHUNK_SIZE):
-            request = api_pb2.FunctionPutInputsRequest(
-                function_id=function_id, inputs=items, function_call_id=function_call_id
-            )
-            logger.debug(
-                f"Pushing {len(items)} inputs to server. Num queued inputs awaiting push is {input_queue.qsize()}."
-            )
-            resp = await retry_transient_errors(
-                client.stub.FunctionPutInputs,
-                request,
-                max_retries=None,
-                max_delay=10,
-                additional_status_codes=[Status.RESOURCE_EXHAUSTED],
-            )
-            count_update()
-            for item in resp.inputs:
-                pending_outputs.setdefault(item.input_id, 0)
-            logger.debug(
-                f"Successfully pushed {len(items)} inputs to server. Num queued inputs awaiting push is {input_queue.qsize()}."
-            )
-
-        have_all_inputs = True
-        yield
-
-    async def get_all_outputs():
-        assert client.stub
-        nonlocal num_inputs, num_outputs, have_all_inputs
-        last_entry_id = "0-0"
-        while not have_all_inputs or len(pending_outputs) > len(completed_outputs):
-            request = api_pb2.FunctionGetOutputsRequest(
-                function_call_id=function_call_id,
-                timeout=OUTPUTS_TIMEOUT,
-                last_entry_id=last_entry_id,
-                clear_on_success=False,
-            )
-            response = await retry_transient_errors(
-                client.stub.FunctionGetOutputs,
-                request,
-                max_retries=20,
-                attempt_timeout=OUTPUTS_TIMEOUT + ATTEMPT_TIMEOUT_GRACE_PERIOD,
-            )
-
-            if len(response.outputs) == 0:
-                continue
-
-            last_entry_id = response.last_entry_id
-            for item in response.outputs:
-                pending_outputs.setdefault(item.input_id, 0)
-                if item.input_id in completed_outputs:
-                    # If this input is already completed, it means the output has already been
-                    # processed and was received again due to a duplicate.
-                    continue
-                completed_outputs.add(item.input_id)
-                num_outputs += 1
-                yield item
-
-    async def get_all_outputs_and_clean_up():
-        assert client.stub
-        try:
-            async for item in get_all_outputs():
-                yield item
-        finally:
-            # "ack" that we have all outputs we are interested in and let backend clear results
-            request = api_pb2.FunctionGetOutputsRequest(
-                function_call_id=function_call_id,
-                timeout=0,
-                last_entry_id="0-0",
-                clear_on_success=True,
-            )
-            await retry_transient_errors(client.stub.FunctionGetOutputs, request)
-
-    async def fetch_output(item: api_pb2.FunctionGetOutputsItem) -> Tuple[int, Any]:
-        try:
-            output = await _process_result(item.result, item.data_format, client.stub, client)
-        except Exception as e:
-            if return_exceptions:
-                output = e
-            else:
-                raise e
-        return (item.idx, output)
-
-    async def poll_outputs():
-        outputs = stream.iterate(get_all_outputs_and_clean_up())
-        outputs_fetched = outputs | pipe.map(fetch_output, ordered=True, task_limit=BLOB_MAX_PARALLELISM)  # type: ignore
-
-        # map to store out-of-order outputs received
-        received_outputs = {}
-        output_idx = 0
-
-        async with outputs_fetched.stream() as streamer:
-            async for idx, output in streamer:
-                count_update()
-                if not order_outputs:
-                    yield _OutputValue(output)
-                else:
-                    # hold on to outputs for function maps, so we can reorder them correctly.
-                    received_outputs[idx] = output
-                    while output_idx in received_outputs:
-                        output = received_outputs.pop(output_idx)
-                        yield _OutputValue(output)
-                        output_idx += 1
-
-        assert len(received_outputs) == 0
-
-    response_gen = stream.merge(drain_input_generator(), pump_inputs(), poll_outputs())
-
-    async with response_gen.stream() as streamer:
-        async for response in streamer:
-            if response is not None:
-                yield response.value
 
 
 # Wrapper type for api_pb2.FunctionStats
@@ -1636,56 +1331,3 @@ async def _gather(*function_calls: _FunctionCall):
 
 
 gather = synchronize_api(_gather)
-
-
-_current_input_id: ContextVar = ContextVar("_current_input_id")
-_current_function_call_id: ContextVar = ContextVar("_current_function_call_id")
-
-
-def current_input_id() -> Optional[str]:
-    """Returns the input ID for the current input.
-
-    Can only be called from Modal function (i.e. in a container context).
-
-    ```python
-    from modal import current_input_id
-
-    @stub.function()
-    def process_stuff():
-        print(f"Starting to process {current_input_id()}")
-    ```
-    """
-    try:
-        return _current_input_id.get()
-    except LookupError:
-        return None
-
-
-def current_function_call_id() -> Optional[str]:
-    """Returns the function call ID for the current input.
-
-    Can only be called from Modal function (i.e. in a container context).
-
-    ```python
-    from modal import current_function_call_id
-
-    @stub.function()
-    def process_stuff():
-        print(f"Starting to process input from {current_function_call_id()}")
-    ```
-    """
-    try:
-        return _current_function_call_id.get()
-    except LookupError:
-        return None
-
-
-def _set_current_context_ids(input_id: str, function_call_id: str) -> Callable[[], None]:
-    input_token = _current_input_id.set(input_id)
-    function_call_token = _current_function_call_id.set(function_call_id)
-
-    def _reset_current_context_ids():
-        _current_input_id.reset(input_token)
-        _current_function_call_id.reset(function_call_token)
-
-    return _reset_current_context_ids
