+++ modal/parallel_map.py
@@ -1,0 +1,224 @@
+import asyncio
+import typing
+from dataclasses import dataclass
+from typing import Any, Callable, Dict, Optional, Set, Tuple
+
+from aiostream import pipe, stream
+from grpclib import Status
+
+from modal._utils.async_utils import queue_batch_iterator, synchronize_api, synchronizer
+from modal._utils.blob_utils import BLOB_MAX_PARALLELISM
+from modal._utils.function_utils import current_input_id
+from modal._utils.grpc_utils import retry_transient_errors
+from modal.config import logger
+from modal.functions import ATTEMPT_TIMEOUT_GRACE_PERIOD, OUTPUTS_TIMEOUT, _create_input, _process_result
+from modal_proto import api_pb2
+
+if typing.TYPE_CHECKING:
+    from modal.client import _Client
+
+
+class _SynchronizedQueue:
+    """mdmd:hidden"""
+
+    # small wrapper around asyncio.Queue to make it cross-thread compatible through synchronicity
+    async def init(self):
+        # in Python 3.8 the asyncio.Queue is bound to the event loop on creation
+        # so it needs to be created in a synchronicity-wrapped init method
+        self.q = asyncio.Queue()
+
+    @synchronizer.no_io_translation
+    async def put(self, item):
+        await self.q.put(item)
+
+    @synchronizer.no_io_translation
+    async def get(self):
+        return await self.q.get()
+
+
+SynchronizedQueue = synchronize_api(_SynchronizedQueue)
+
+
+@dataclass
+class _OutputValue:
+    # box class for distinguishing None results from non-existing/None markers
+    value: Any
+
+
+MAP_INVOCATION_CHUNK_SIZE = 49
+
+
+async def _map_invocation(
+    function_id: str,
+    raw_input_queue: _SynchronizedQueue,
+    client: "_Client",
+    order_outputs: bool,
+    return_exceptions: bool,
+    count_update_callback: Optional[Callable[[int, int], None]],
+):
+    assert client.stub
+    request = api_pb2.FunctionMapRequest(
+        function_id=function_id,
+        parent_input_id=current_input_id() or "",
+        function_call_type=api_pb2.FUNCTION_CALL_TYPE_MAP,
+        return_exceptions=return_exceptions,
+    )
+    response = await retry_transient_errors(client.stub.FunctionMap, request)
+
+    function_call_id = response.function_call_id
+
+    have_all_inputs = False
+    num_inputs = 0
+    num_outputs = 0
+
+    def count_update():
+        if count_update_callback is not None:
+            count_update_callback(num_outputs, num_inputs)
+
+    pending_outputs: Dict[str, int] = {}  # Map input_id -> next expected gen_index value
+    completed_outputs: Set[str] = set()  # Set of input_ids whose outputs are complete (expecting no more values)
+
+    input_queue: asyncio.Queue = asyncio.Queue()
+
+    async def create_input(argskwargs):
+        nonlocal num_inputs
+        idx = num_inputs
+        num_inputs += 1
+        (args, kwargs) = argskwargs
+        return await _create_input(args, kwargs, client, idx)
+
+    async def input_iter():
+        while 1:
+            raw_input = await raw_input_queue.get()
+            if raw_input is None:  # end of input sentinel
+                return
+            yield raw_input  # args, kwargs
+
+    async def drain_input_generator():
+        # Parallelize uploading blobs
+        proto_input_stream = stream.iterate(input_iter()) | pipe.map(
+            create_input,  # type: ignore[reportArgumentType]
+            ordered=True,
+            task_limit=BLOB_MAX_PARALLELISM,
+        )
+        async with proto_input_stream.stream() as streamer:
+            async for item in streamer:
+                await input_queue.put(item)
+
+        # close queue iterator
+        await input_queue.put(None)
+        yield
+
+    async def pump_inputs():
+        assert client.stub
+        nonlocal have_all_inputs, num_inputs
+        async for items in queue_batch_iterator(input_queue, MAP_INVOCATION_CHUNK_SIZE):
+            request = api_pb2.FunctionPutInputsRequest(
+                function_id=function_id, inputs=items, function_call_id=function_call_id
+            )
+            logger.debug(
+                f"Pushing {len(items)} inputs to server. Num queued inputs awaiting push is {input_queue.qsize()}."
+            )
+            resp = await retry_transient_errors(
+                client.stub.FunctionPutInputs,
+                request,
+                max_retries=None,
+                max_delay=10,
+                additional_status_codes=[Status.RESOURCE_EXHAUSTED],
+            )
+            count_update()
+            for item in resp.inputs:
+                pending_outputs.setdefault(item.input_id, 0)
+            logger.debug(
+                f"Successfully pushed {len(items)} inputs to server. Num queued inputs awaiting push is {input_queue.qsize()}."
+            )
+
+        have_all_inputs = True
+        yield
+
+    async def get_all_outputs():
+        assert client.stub
+        nonlocal num_inputs, num_outputs, have_all_inputs
+        last_entry_id = "0-0"
+        while not have_all_inputs or len(pending_outputs) > len(completed_outputs):
+            request = api_pb2.FunctionGetOutputsRequest(
+                function_call_id=function_call_id,
+                timeout=OUTPUTS_TIMEOUT,
+                last_entry_id=last_entry_id,
+                clear_on_success=False,
+            )
+            response = await retry_transient_errors(
+                client.stub.FunctionGetOutputs,
+                request,
+                max_retries=20,
+                attempt_timeout=OUTPUTS_TIMEOUT + ATTEMPT_TIMEOUT_GRACE_PERIOD,
+            )
+
+            if len(response.outputs) == 0:
+                continue
+
+            last_entry_id = response.last_entry_id
+            for item in response.outputs:
+                pending_outputs.setdefault(item.input_id, 0)
+                if item.input_id in completed_outputs:
+                    # If this input is already completed, it means the output has already been
+                    # processed and was received again due to a duplicate.
+                    continue
+                completed_outputs.add(item.input_id)
+                num_outputs += 1
+                yield item
+
+    async def get_all_outputs_and_clean_up():
+        assert client.stub
+        try:
+            async for item in get_all_outputs():
+                yield item
+        finally:
+            # "ack" that we have all outputs we are interested in and let backend clear results
+            request = api_pb2.FunctionGetOutputsRequest(
+                function_call_id=function_call_id,
+                timeout=0,
+                last_entry_id="0-0",
+                clear_on_success=True,
+            )
+            await retry_transient_errors(client.stub.FunctionGetOutputs, request)
+
+    async def fetch_output(item: api_pb2.FunctionGetOutputsItem) -> Tuple[int, Any]:
+        try:
+            output = await _process_result(item.result, item.data_format, client.stub, client)
+        except Exception as e:
+            if return_exceptions:
+                output = e
+            else:
+                raise e
+        return (item.idx, output)
+
+    async def poll_outputs():
+        outputs = stream.iterate(get_all_outputs_and_clean_up())
+        outputs_fetched = outputs | pipe.map(fetch_output, ordered=True, task_limit=BLOB_MAX_PARALLELISM)  # type: ignore
+
+        # map to store out-of-order outputs received
+        received_outputs = {}
+        output_idx = 0
+
+        async with outputs_fetched.stream() as streamer:
+            async for idx, output in streamer:
+                count_update()
+                if not order_outputs:
+                    yield _OutputValue(output)
+                else:
+                    # hold on to outputs for function maps, so we can reorder them correctly.
+                    received_outputs[idx] = output
+                    while output_idx in received_outputs:
+                        output = received_outputs.pop(output_idx)
+                        yield _OutputValue(output)
+                        output_idx += 1
+
+        assert len(received_outputs) == 0
+
+    response_gen = stream.merge(drain_input_generator(), pump_inputs(), poll_outputs())
+
+    async with response_gen.stream() as streamer:
+        async for response in streamer:
+            if response is not None:
+                yield response.value
